Permutation Test for Cluster Properties

This function compares the properties of actual clusters to properties of randomly permuted clusters which retain the original distribution of patient types (index start, index not start, convert) within them.

Usage

cluster_property_perm_test(
  clusters,
  pt_trace,
  seq2pt,
  ip_pt_seqs,
  ip_seqs,
  dates,
  snp_dist,
  nperm = 1000,
  floor_trace = NULL,
  room_trace = NULL,
  num_cores = detectCores() - 1
)

Arguments

clusters

A named numeric vector where names are sequence IDs and values are subtrees defining the cluster.

pt_trace

A matrix or data frame with rows representing days and columns representing patients.

seq2pt

A named vector mapping sequence IDs to patient IDs.

ip_pt_seqs

A vector of sequence IDs corresponding to intake positive patients.

ip_seqs

A vector of sequence IDs corresponding to intake patient sequences presumed to be imported.

dates

A vector of isolate dates named by sequence IDs.

snp_dist

A matrix of SNP distances between isolates.

nperm

The number of permutations to perform (default is 1000).

floor_trace

An optional floor trace (rows: days, columns: patients).

room_trace

An optional room trace (rows: days, columns: patients).

num_cores

The number of cores to use for parallel processing (default is one less than the total number of cores). This is only for *nix systems and does not work on Windows - for Windows, set num_cores = 1.

Value

A matrix of permutation statistics. Rows correspond to each permutation (with the last row containing the observed statistic) and columns correspond to the computed properties.