Get Transmission Clusters

This function identifies transmission clusters based on the number of shared variants.

Usage

get_tn_clusters_MSV_SVst_index_first(
  dna_aln,
  snp_dist,
  ip_seqs,
  ip_pt_seqs,
  seq2pt,
  dates,
  method = c("nj", "pars"),
  tree = NULL
)

Arguments

dna_aln

A DNA object for sequences of interest.

snp_dist

A matrix of SNP distances between isolates constructed using a model of DNA evolution.

ip_seqs

A vector of sequence IDs which correspond to intake patient sequences presumed to be imported.

ip_pt_seqs

A vector of sequence IDs which correspond to intake-positive patients.

seq2pt

A named vector linking sequence IDs to patient IDs.

dates

A vector of isolate dates named by sequence IDs.

method

A string indicating the method to use for tree construction. Options are "nj" (neighbor-joining) or "pars" (maximum parsimony). Default is "pars".

tree

A tree to use if provided instead of the default neighbor joining tree.

Value

A vector indicating the cluster that each isolate belongs to.

Details

Clustering is performed to identify the maximal clusters containing a single intake-positive patient that occurs before all cluster converts. The clustering metric is the number of shared variants, and clusters can have multiple intake-positive patients if they share an identical number of variants with other cluster members or intake-positive patients occur after converts. This clustering also requires that clusters be defined by at least one shared variant that other isolates don't have.